Update color uniform project to color palette
