+++ Assignments/Assignment-1/ColorPalette/cvec.h
