+++ Assignments/Assignment-1/ColorPalette/glsupport.cpp
