+++ Assignments/Assignment-1/ColorPalette/glsupport.h
