+++ Assignments/Assignment-1/ColorPalette/stb_image.h
