+++ Assignments/Assignment-1/ColorPalette/superman.png
