+++ Assignments/Assignment-1/ColorPalette/vertex.glsl
@@ -1,4 +1,5 @@
 attribute vec4 positionAttribute;
+varying vec2 varyingPosiiton;
 uniform vec2 positionUniform;
 
 void main() {
