+++ Assignments/ColorPalette.xcodeproj/project.xcworkspace/contents.xcworkspacedata
