+++ Assignments/ColorPalette.xcodeproj/xcuserdata/nandukalidindi.xcuserdatad/xcschemes/xcschememanagement.plist
